Updated DebugVoucherStepDrawer and updated tests and snapshots
